chore: clear remaining clippy warnings in the lib target
- turn the receipt formatter header into a module doc comment
- use clamp for the dashboard and product-shortcut limit bounds
- swap map_or for is_some_and on the pinned-shortcut lookup
- pass AppError::Internal directly instead of wrapping it in a closure
- drop the no-op i64 casts on refund and exchange quantities
- allow too_many_arguments on the two PPOB commands whose flat parameter
  lists are the IPC wire shape, with the reason recorded inline

+++ src-tauri/src/services/refunds/create.rs
@@ -255,7 +255,7 @@ pub async fn create(
                 DbBackend::Sqlite,
                 "UPDATE products SET stock = stock + $1, updated_at = $2 WHERE id = $3",
                 vec![
-                    (item_input.quantity as i64).into(),
+                    item_input.quantity.into(),
                     now.clone().into(),
                     product_id.into(),
                 ],
@@ -371,7 +371,7 @@ pub async fn create(
                 DbBackend::Sqlite,
                 "UPDATE products SET stock = stock - $1, updated_at = $2 WHERE id = $3",
                 vec![
-                    (ei_input.quantity as i64).into(),
+                    ei_input.quantity.into(),
                     now.clone().into(),
                     ei_input.product_id.into(),
                 ],
